requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==3.2.3
django-ckeditor==6.0.0
django-import-export==2.5.0
django-js-asset==1.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
